Finished moving models to one place
